Initial commit
This is a copy of the directory that we're currently hosting on our old
Linode instance. We've got the original caffeine monitor Rails app in
the Archives folder in Dropbox.

I'm planning to add to this repo so that I can use recap to deploy it
to our new Linode instance.